@@ -8,9 +8,7 @@ use Symfony\Component\Routing\Annotation\Route;
 
 class MainController extends AbstractController
 {
-    /**
-     * @Route("/", name="app_main")
-     */
+    #[Route(path: "/", name: "app_main", methods: ['GET'])]
     public function index(): Response
     {
         return $this->render('main/index.html.twig', [
@@ -18,13 +16,11 @@ class MainController extends AbstractController
         ]);
     }
 
-    /**
-     * @Route("/folder/", name="app_folder")
-     */
+    #[Route(path: "/folder/", name: "app_folder", methods: ['GET'])]
     public function folder(): Response
     {
         return $this->render('main/index.html.twig', [
-            'controller_name' => 'MainController',
+            'controller_name' => 'FolderController',
         ]);
     }
 }
