@@ -9,6 +9,7 @@ use Sonata\AdminBundle\Datagrid\DatagridMapper;
 use Sonata\AdminBundle\Datagrid\ListMapper;
 use Sonata\AdminBundle\Form\FormMapper;
 use Sonata\AdminBundle\Show\ShowMapper;
+use Sonata\Form\Type\DateTimePickerType;
 use Sonata\MediaBundle\Form\Type\MediaType;
 
 final class ArticleAdmin extends AbstractAdmin
@@ -26,8 +27,8 @@ final class ArticleAdmin extends AbstractAdmin
     protected function configureListFields(ListMapper $list): void
     {
         $list
-            ->add('id')
-            ->add('name', null, ['editable' => true], ['editable' => true])
+            ->addIdentifier('id')
+            ->add('name', null, ['editable' => true])
             ->add('datePublication')
             ->add(ListMapper::NAME_ACTIONS, null, [
                 'actions' => [
